Add Event model with fields for sport, location, date, maxParticipants, and hostId; remove unused Event definition from User model and updated sprint 2
