+++ Sprint2.md
@@ -10,3 +10,25 @@ router/controller
 API testing: post form , and a get from data base
 front end:
 cards to get from data base and show on app
+
+
+
+DONE
+
+Authentication System
+
+User registration
+Login/Logout functionality
+Protected routes using middleware
+Session management
+Basic UI Components
+
+Navbar with navigation
+Login/Register page
+Home page layout
+New game submission form
+Database Setup
+
+MySQL connection
+User model with basic fields
+Event model structure
