+++ api/model/user.js
@@ -23,30 +23,5 @@ const User = sequelize.define('user', {
     }
 })
 
-const Event = sequelize.define('event', {
-    id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true
-    },
-    username: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        reservation_time: DataTypes.TIME
-    },
-    address: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    number_of_people: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    reservation_date: {
-        type: DataTypes.DATE,
-        allowNull: false
-    }
-}
-)
 
-export { User, Event };
+export default User
